@@ -23,87 +23,96 @@
     "clean": "rimraf dist.dev dist.prod"
   },
   "dependencies": {
-    "@babel/polyfill": "^7.0.0",
-    "@vimeo/player": "2.6.3",
+    "@babel/polyfill": "^7.2.5",
+    "@vimeo/player": "2.6.5",
     "axios": "0.18.0",
     "d3": "^5.7.0",
-    "formik": "^1.3.1",
+    "formik": "^1.4.1",
     "lodash": "4.17.11",
-    "react": "16.5.2",
-    "react-dom": "16.5.2",
+    "react": "16.7.0",
+    "react-dom": "16.7.0",
     "react-facebook-login": "^4.1.1",
-    "react-redux": "5.0.7",
+    "react-redux": "6.0.0",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
+    "react-select": "^2.2.0",
     "reactstrap": "^6.5.0",
-    "redux": "4.0.0",
-    "redux-form": "^7.4.2",
-    "redux-saga": "0.16.1",
+    "redux": "4.0.1",
+    "redux-actions": "^2.6.4",
+    "redux-form": "^8.1.0",
+    "redux-saga": "0.16.2",
     "yup": "^0.26.6"
   },
   "devDependencies": {
-    "@babel/cli": "^7.1.2",
-    "@babel/core": "^7.1.2",
-    "@babel/plugin-proposal-class-properties": "^7.1.0",
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.2.2",
+    "@babel/plugin-proposal-class-properties": "^7.2.3",
     "@babel/plugin-transform-modules-commonjs": "^7.1.0",
-    "@babel/preset-env": "^7.1.0",
+    "@babel/preset-env": "^7.2.3",
     "@babel/preset-react": "^7.0.0",
     "@babel/register": "^7.0.0",
-    "autoprefixer": "9.1.5",
+    "autoprefixer": "^9.4.3",
     "babel-core": "^7.0.0-bridge.0",
+    "babel-eslint": "^10.0.1",
     "babel-jest": "^23.6.0",
     "babel-loader": "^8.0.4",
     "case-sensitive-paths-webpack-plugin": "2.1.2",
     "chalk": "2.4.1",
     "cors": "^2.8.4",
     "cross-env": "^5.2.0",
-    "css-loader": "^1.0.0",
-    "dotenv": "6.0.0",
+    "css-loader": "^2.1.0",
+    "dotenv": "6.2.0",
     "dotenv-expand": "4.2.0",
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-16": "^1.1.1",
-    "eslint": "5.6.1",
-    "eslint-config-react-app": "^3.0.3",
-    "eslint-loader": "2.1.1",
-    "eslint-plugin-flowtype": "2.50.3",
+    "eslint": "5.11.1",
+    "eslint-config-react-app": "^3.0.6",
+    "eslint-loader": "^2.1.1",
+    "eslint-plugin-flowtype": "3.2.0",
     "eslint-plugin-import": "2.14.0",
     "eslint-plugin-jsx-a11y": "6.1.2",
-    "eslint-plugin-react": "7.11.1",
+    "eslint-plugin-react": "^7.12.0",
     "extract-text-webpack-plugin": "3.0.2",
-    "file-loader": "^2.0.0",
+    "file-loader": "^3.0.1",
     "font-awesome": "4.7.x",
-    "fs-extra": "7.0.0",
+    "fs-extra": "7.0.1",
     "html-webpack-plugin": "3.2.0",
+    "identity-obj-proxy": "^3.0.0",
     "jest": "^23.6.0",
-    "mini-css-extract-plugin": "^0.4.1",
+    "mini-css-extract-plugin": "^0.5.0",
+    "node-noop": "^1.0.0",
     "node-sass": "^4.9.2",
     "object-assign": "4.1.1",
-    "postcss-flexbugs-fixes": "4.1.0",
+    "postcss-flexbugs-fixes": "^4.1.0",
     "postcss-loader": "^3.0.0",
-    "precss": "^3.1.2",
+    "precss": "^4.0.0",
     "promise": "8.0.2",
-    "raf": "3.4.0",
-    "react-dev-utils": "^6.0.4",
-    "react-hot-loader": "^4.3.5",
-    "react-test-renderer": "^16.5.2",
-    "regenerator-runtime": "^0.12.1",
-    "resolve": "1.8.1",
+    "raf": "3.4.1",
+    "react-dev-utils": "^7.0.0",
+    "react-hot-loader": "^4.6.3",
+    "react-test-renderer": "^16.7.0",
+    "regenerator-runtime": "^0.13.1",
+    "resolve": "1.9.0",
     "sass-loader": "^7.0.1",
-    "sinon": "^6.3.5",
-    "style-loader": "0.23.0",
+    "sinon": "^7.2.2",
+    "style-loader": "0.23.1",
     "sw-precache-webpack-plugin": "0.11.5",
-    "url-loader": "1.1.1",
-    "webpack": "^4.8.3",
+    "url-loader": "1.1.2",
+    "wallaby-webpack": "^3.9.13",
+    "webpack": "^4.28.2",
     "webpack-bundle-tracker": "^0.3.0",
     "webpack-cli": "^3.1.2",
-    "webpack-dev-server": "^3.1.4",
+    "webpack-dev-server": "^3.1.14",
     "webpack-merge": "^4.1.2"
   },
   "jest": {
-    "setupTestFrameworkScriptFile": "./node_modules/jest-enzyme/lib/index.js",
-    "unmockedModulePathPatterns": [
-      "react",
-      "enzyme",
-      "jest-enzyme"
-    ],
-    "testURL": "http://localhost/"
+    "testURL": "http://localhost/",
+    "moduleNameMapper": {
+      "~/(.*)": "<rootDir>/src/$1",
+      "\\.(css|scss)$": "identity-obj-proxy"
+    },
+    "transform": {
+      "\\.(scss|css)$": "<rootDir>/fileTransformer.js"
+    }
   }
 }
