@@ -1,44 +1,34 @@
-from rest_framework.request import Request
-from .apiutils import get_current_season
+from rest_framework.views import APIView
+from rest_framework.response import Response
 
 from TBLib.manager import TeamManager
-from TBLib.teamgen.DBTeams import DBTeams
 
 from .apiutils import JSONResponse
 
 
-def pick_teams(request, date=None):
-    """
-    """
-    r = Request(request)
+class Teams(APIView):
 
-    if r.method == 'POST':
-        print("pick_teams POST for date %s" % date)
-        matchData = {}
-        season = get_current_season()
-        if date and season:
-            dbTeams = DBTeams()
-
-            men, women = dbTeams.get_players(date)
-            mgr = TeamManager()
-            mgr.pick_teams(men=men, women=women)
-
-            matchData = mgr.query_match(date)
-
-        return JSONResponse({'status': 'POST Done', 'date': date, 'teams': matchData})
-
-
-def query_teams(request, date=None):
-    """
-    """
-    r = Request(request)
-
-    if r.method == 'GET':
+    def get(self, request, date=None):
         print("pick_teams GET for date %s" % date)
-        matchData = {}
-        if date:
-            mgr = TeamManager()
 
-            matchData = mgr.query_match(date)
+        mgr = TeamManager()
+        match_data = mgr.query_match(date)
 
-        return JSONResponse({'status': 'GET Done', 'date': date, 'teams': matchData})
+        return Response({
+            'status': 'GET Done',
+            'date': date,
+            'teams': match_data
+        })
+
+    def post(self, request, date=None):
+        # Date can be part of the URL, or the post data.
+        date = request.data.get('date', date)
+        iterations = request.data.get('iterations', 10)
+        tries = request.data.get('tries', 5)
+
+        mgr = TeamManager()
+        result = mgr.pick_teams_for_date(date,
+                                         iterations=iterations,
+                                         max_tries=tries)
+
+        return Response(result)
