@@ -4,11 +4,11 @@ from .availability import AvailabilityView
 from .members import SeasonPlayerView
 from .blocksheet import blockSheet
 from .blockschedule import (BlockNotifyer, getBlockDates,
-                            getSubList, blockPlayers, blockSchedule,
-                            getMatchData)
+                            getSubList, blockPlayers,
+                            getMatchData, BlockSchedule)
 
 from .views import get_seasons, get_current_season, get_latest_buzz
-from .teams import pick_teams
+from .teams import Teams
 
 from . import converters
 
@@ -19,8 +19,8 @@ urlpatterns = (
     path('seasons', get_seasons),
     path('currentseason', get_current_season),
     path('buzz', get_latest_buzz),
-    path('pickteams/<date:date>', pick_teams),
-    path('pickteams', pick_teams),
+    path('pickteams/<date:date>', Teams.as_view()),
+    path('pickteams', Teams.as_view()),
     path('blocksheet/<date:date>', blockSheet),
     path('blocksheet', blockSheet),
     path('blockdates', getBlockDates),
@@ -28,11 +28,12 @@ urlpatterns = (
     path('subs', getSubList),
     path('blockplayers/<date:date>', blockPlayers),
     path('blockplayers', blockPlayers),
-    path('blockschedule/<date:date>', blockSchedule),
-    path('blockschedule', blockSchedule),
+    path('blockschedule/', BlockSchedule.as_view()),
+    path('blockschedule/<date:date>', BlockSchedule.as_view()),
     path('matchdata/<date:date>', getMatchData),
     path('matchdata', getMatchData),
     path('availability', AvailabilityView.as_view()),
+    path('availability/', AvailabilityView.as_view()),
     path('members/<int:id>', SeasonPlayerView.as_view()),
     path('members', SeasonPlayerView.as_view()),
     path('schedule/notify/<date:date>', BlockNotifyer.as_view(), name="schedule_notify"),
