@@ -1,7 +1,9 @@
-import mkStore from './make_store'
-import {selectors} from './selectors'
+import mkStore from '~/Schedule/modules/make_store'
 
-export {selectors}
+import { selectors } from './selectors'
+export { selectors }
+
+import { actions } from './actions'
+export { actions }
 
 export default mkStore
-
