@@ -1,6 +1,6 @@
 import reducer from './reducers'
 
-import {NAME} from './constants'
+import {MODULE_NAME} from './constants'
 import * as types from './constants'
 import * as actions from './actions'
 import * as selectors from './selectors'
@@ -8,6 +8,6 @@ import rootSaga from './sagas'
 
 import {APP_NAME} from './constants'
 export {APP_NAME}
-export {types, actions, selectors, NAME, rootSaga};
+export {types, actions, selectors, MODULE_NAME, rootSaga};
 
 export default reducer
