@@ -1,8 +1,4 @@
-// If a feature has more facets, you should definitely use multiple
-// reducers to handle different parts of the state shape.
-// Additionally, don’t be afraid to use combineReducers as much as needed.
-// This gives you a lot of flexibility when working with a complex state shape.
-import {combineReducers} from 'redux'
+import { handleActions } from 'redux-actions'
 import * as types from './constants'
 
 const initialState = {
@@ -20,13 +16,79 @@ const initialState = {
     players_by_id: {},
     guys: [],
     gals: [],
-  }
+  },
+  originalCouples: [],
+  blockdates: [],
+  blockplayers: {guys:[], gals:[], couples: []},
+  subs: {},
 }
 
-const mainReducer = (state = initialState, action) => {
-  switch(action.type) {
-    default: return state;
-  }
+const filter_sub = (sub) => {
+  const {name, id} = sub
+  return {name, id}
+}
+const filter_player = (p) => {
+  const {name, id} = p
+  return {name, id}
 }
 
-export default mainReducer
+const reducer = handleActions(
+    {
+        [types.SET_BLOCKDATES](state, { payload }) {
+          const blockdates = payload
+          return {
+            ...state,
+            blockdates
+          }
+        },
+
+        [types.UPDATE_PLAY_SCHEDULE](state, { payload }) {
+          return {...state}
+        },
+
+        [types.FETCH_BLOCK_PLAYERS_SUCCEED](state, { payload }) {
+          const blockplayers = payload
+          const { couples, guys, gals } = blockplayers
+
+          return {
+            ...state,
+            blockplayers,
+            originalCouples: [...couples],
+            guys: guys.map(filter_player),
+            gals: gals.map(filter_player),
+          }
+        },
+
+        [types.GET_SUBS](state, { payload }) {
+          const subs = payload
+          const {guysubs: guys, galsubs: gals} = payload
+          return {
+            ...state,
+            subs,
+            subs_guys: guys.map(filter_sub),
+            subs_gals: gals.map(filter_sub),
+          }
+        },
+
+        [types.BLOCK_PLAYER_CHANGED](state, { payload }) {
+          const selectedPlayer = payload
+          const { key, gender, player } = selectedPlayer
+          const currentPlayer = state.blockplayers.couples[key][gender]
+
+          state.blockplayers.couples.splice(key, 1, {
+            ...state.blockplayers.couples[key],
+            [gender]: player
+          })
+
+          const subIndex = state.subs[`${gender}subs`].findIndex(sub => sub.id === selectedPlayer.player.id)
+          state.subs[`${gender}subs`].splice(subIndex, 1, currentPlayer)
+
+          return { ...state }
+        },
+
+
+    },
+    initialState
+)
+
+export default reducer
