@@ -2,8 +2,7 @@ import axios from 'axios'
 
 const instance = axios.create({
   // ToDo: need a solution that works for deployed app.
-  baseURL: 'https://tennisblock.local:8002'
+  baseURL: `${window.location.protocol}//${window.location.host}`
 })
 
 export default instance
-
