@@ -1,4 +1,3 @@
-const APP_NAME = 'tennisblock'
+export const APP_NAME = 'tennisblock'
 
-export {APP_NAME}
 
